requirements.txt
streamlit>=1.30
numpy>=1.23
pillow>=10.0
pywavelets>=1.5
mpmath>=1.3
scipy>=1.10
scikit-image>=0.21
